requirements.txt
fastapi==0.115.9
numpy==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
slowapi==0.1.9
SQLAlchemy==2.0.38
tensorflow==2.18.0
tensorflow_hub==0.16.1
torch==2.6.0
transformers==4.49.0
uvicorn==0.34.0

